Revert "centralize accelerometer impl"
This reverts commit 8a8c508072635dc1de362207e651ca8fce31978d.
